[2741] edited kdocs, improved naming and code formatting as requested in the PR review
