+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/selectedmessage/SelectedMessageOptions.kt
@@ -13,46 +13,43 @@ import io.getstream.chat.android.client.models.Reaction
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.common.state.MessageAction
 import io.getstream.chat.android.common.state.React
-import io.getstream.chat.android.compose.state.messages.list.MessageOptionState
+import io.getstream.chat.android.compose.state.messages.list.MessageOptionItemState
 import io.getstream.chat.android.compose.ui.components.messageoptions.MessageOptions
 import io.getstream.chat.android.compose.ui.components.messageoptions.defaultMessageOptionsState
 import io.getstream.chat.android.compose.ui.components.reactionoptions.ReactionOptions
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 
 /**
- * Uses slots [headerContent] and [bodyContent] to display it's content.
- * By default shows [ReactionOptions] as header content and [MessageOptions] as body content.
+ * Represents the options user can take after selecting a message.
  *
- * @param message Used to gather the necessary message information (such as ownership, id, etc.) in order to properly display [SelectedMessageOptions].
- * @param modifier Compose UI [Modifier] that is applied to the internally used [Column].
- * @param reactionTypes By default used to display all the available reaction options inside of [headerContent].
- * @param messageOptions By default used to display all available message options inside of [bodyContent].
- * @param onMessageAction By default a Handler used to propagate click events on individual reaction and message option elements
- * inside of the header and body content.
- * @param headerContent Leading vertical Composable slot
- * @param bodyContent Trailing vertical Composable slot.
+ * @param message The selected message.
+ * @param messageOptions The available message options within the menu.
+ * @param onMessageAction Handler that propagates click events on each item.
+ * @param modifier Modifier for styling.
+ * @param reactionTypes The available reactions within the menu.
+ * @param headerContent Leading vertical Composable that allows the user to customize the content shown in [SelectedMessageOptions].
+ * By default shows reaction options.
+ * @param bodyContent Trailing vertical Composable that allows the user to customize the content shown in [SelectedMessageOptions].
+ * By Default shows message options.
  */
 @Composable
 public fun SelectedMessageOptions(
     message: Message,
+    messageOptions: List<MessageOptionItemState>,
+    onMessageAction: (MessageAction) -> Unit,
     modifier: Modifier = Modifier,
     reactionTypes: Map<String, Int> = ChatTheme.reactionTypes,
-    messageOptions: List<MessageOptionState>,
-    onMessageAction: (MessageAction) -> Unit,
     headerContent: @Composable ColumnScope.() -> Unit = {
         ReactionOptions(
             modifier = Modifier
                 .fillMaxWidth()
                 .padding(horizontal = 16.dp, vertical = 12.dp),
             reactionTypes = reactionTypes,
-            onReactionOptionClicked = {
+            onReactionOptionSelected = {
                 onMessageAction(
                     React(
-                        Reaction(
-                            messageId = message.id,
-                            type = it.type,
-                        ),
-                        message
+                        reaction = Reaction(messageId = message.id, type = it.type),
+                        message = message
                     )
                 )
             },
@@ -61,8 +58,8 @@ public fun SelectedMessageOptions(
     },
     bodyContent: @Composable ColumnScope.() -> Unit = {
         MessageOptions(
-            messageOptionStateList = messageOptions,
-            onMessageItemOptionClicked = {
+            options = messageOptions,
+            onMessageOptionSelected = {
                 onMessageAction(it.action)
             }
         )
